[NEW] Swap History and Dashboard (#65)
* swapped history and dashboard

* fixed loading issue

* fixed modal css
